requirements.txt
flask
pandas
openpyxl
cryptography
python-dotenv
